agregado de RowAsistencias
